# docs/swagger.yaml
...
info:
  contact: {}
  description: JetKVM backend API documentation
  title: JetKVM API
  version: "1.0"
paths:
...
  /storage/upload:
    post:
      consumes:
      - application/octet-stream
      description: Upload a file to the device storage using HTTP POST
      parameters:
      - description: Upload session ID
        in: query
        name: uploadId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AuthToken: []
      summary: Upload file to device storage
      tags:
      - storage
...
securityDefinitions:
  AuthToken:
    in: cookie
    name: authToken
    type: apiKey
swagger: "2.0"